[FEATURE-13] Move api to Shared folder
