+++ src/Host/ClientApp/src/app/index.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import { configureStore } from '@reduxjs/toolkit';
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import { localhostApi } from './localhostApi';
+import { localhostApi } from '../shared/localhostApi';
 import { counterReducer } from '../features/Counter';
 
 // Всякий раз, когда отправляется (dispatched) действие (action), Redux будет обновлять каждое
