+++ src/Host/ClientApp/src/shared/localhostApi.ts
